multi_agent_system/requirements.txt
fastapi
uvicorn
redis
pydantic
pdfplumber
PyPDF2
faker
langchain
requests
python-dotenv
textblob
openai
